+++ tests/PromiseTest/FullTestTrait.php
@@ -10,5 +10,5 @@ trait FullTestTrait
         PromiseRejectedTestTrait,
         ResolveTestTrait,
         RejectTestTrait,
-        ProgressTestTrait;
+        NotifyTestTrait;
 }